Gå direkte mot search, og ikke via suggestion. Fjern gammelt endepunkt
