+++ Digipost.Api.Client.Tests/Integration/DomainUtility.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Security.Policy;
-using ApiClientShared;
+﻿using ApiClientShared;
 using Digipost.Api.Client.Domain;
-using Digipost.Api.Client.Domain.Autocomplete;
 using Digipost.Api.Client.Domain.Enums;
 
 namespace Digipost.Api.Client.Tests.Integration
@@ -25,22 +21,6 @@ namespace Digipost.Api.Client.Tests.Integration
         {
             var identification = new Identification(IdentificationChoice.PersonalidentificationNumber, "00000000000");
             return identification;
-        }
-
-        public static AutocompleteSuggestionResults GetAutocompleteSuggestionResults()
-        {
-            
-            var autocompleteSuggestionList = new List<AutocompleteSuggestion>
-            {
-                GetSuggestion("marit johansen")
-            };
-
-            return new AutocompleteSuggestionResults( ){AutcompleteSuggestions = autocompleteSuggestionList};
-        }
-
-        public static AutocompleteSuggestion GetSuggestion(string searchString = "marit", string name = "marit johansen")
-        {
-            return new AutocompleteSuggestion(searchString, new Link("https://qa2.api.digipost.no/relations/search", "https://qa2.api.digipost.no/recipients/search/" + Uri.EscapeUriString(name), "application/vnd.digipost-v6+xml"));
         }
     }
 }
