+++ Digipost.Api.Client/Api/IDigipostApi.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Digipost.Api.Client.Action;
 using Digipost.Api.Client.Domain;
+using Digipost.Api.Client.Domain.PersonDetails;
 
 namespace Digipost.Api.Client.Api
 {
@@ -8,11 +9,15 @@ namespace Digipost.Api.Client.Api
     {
         MessageDeliveryResult SendMessage(Message message);
 
-        IdentificationResult Identify(Identification identification);
-
         Task<MessageDeliveryResult> SendMessageAsync(Message message);
 
+        IdentificationResult Identify(Identification identification);
+
         Task<IdentificationResult> IdentifyAsync(Identification identification);
+
+        Task<PersonDetailsResult> SearchAsync(string search);
+
+        PersonDetailsResult Search(string search);
 
         IDigipostActionFactory DigipostActionFactory { get; set; }
     }
