+++ Digipost.Api.Client/DigipostClient.cs
@@ -1,8 +1,6 @@
 ﻿using System.Security.Cryptography.X509Certificates;
-using System.Security.Policy;
 using System.Threading.Tasks;
 using Digipost.Api.Client.Domain;
-using Digipost.Api.Client.Domain.Autocomplete;
 using Digipost.Api.Client.Domain.PersonDetails;
 
 namespace Digipost.Api.Client.Api
@@ -43,24 +41,14 @@ namespace Digipost.Api.Client.Api
             return api.SendMessageAsync(message);
         }
 
-        public AutocompleteSuggestionResults Autocomplete(string query)
+        public PersonDetailsResult Search(string query)
         {
-            return api.Autocomplete(query);
+            return api.Search(query);
         }
 
-        public Task<AutocompleteSuggestionResults> AutocompleteAsync(string query)
+        public Task<PersonDetailsResult> SearchAsync(string query)
         {
-            return api.AutocompleteAsync(query);
-        } 
-
-        public PersonDetailsResult GetPersonDetails(AutocompleteSuggestion suggestions)
-        {
-            return api.GetPersonDetails(suggestions);
-        }
-
-        public Task<PersonDetailsResult> GetPersonDetailsAsync(AutocompleteSuggestion suggestions)
-        {
-            return api.GetPersonDetailsAsync(suggestions);
+            return api.SearchAsync(query);
         }
     }
 }
